added front and back end authentication, changed the article page design a bit

+++ client/src/services/ArticlesService.js
@@ -5,22 +5,22 @@ export default {
     return Api.get('api/articles')
   },
   addArticle (params) {
-    return Api.post('api/articles', params)
+    return Api.post('api/articles', params, { headers: { Authorization: localStorage.getItem('token') } })
   },
   updateArticle (params) {
-    return Api.put(`api/articles/${params.id}`, params)
+    return Api.put(`api/articles/${params.id}`, params, { headers: { Authorization: localStorage.getItem('token') } })
   },
   uploadArticleImage (params) {
     let form = new FormData()
 
     form.append('articleImg', params.image)
 
-    return Api.patch(`api/articles/${params.id}/image`, form)
+    return Api.patch(`api/articles/${params.id}/image`, form, { headers: { Authorization: localStorage.getItem('token') } })
   },
   getArticle (params) {
-    return Api.get(`api/article/${params.id}`)
+    return Api.get(`api/article/${params.id}`, { headers: { Authorization: localStorage.getItem('token') } })
   },
   deleteArticle (id) {
-    return Api.delete(`api/articles/${id}`)
+    return Api.delete(`api/articles/${id}`, { headers: { Authorization: localStorage.getItem('token') } })
   }
 }
